+++ Docs/CHANGELOG.md
@@ -4,6 +4,10 @@ Changes are relative to v1.0.0
 
 ## Changes since v1.1.0
 
+### Extended Features
+
+- Improve error handling, specifically when a discovery timeout occurs it is now logged as a warning at logging level normal.
+
 ### Bug fixes
 
 - Bug: CHECK_FALSE and REQUIRE_FALSE failed output expansion shows '!' in front of value results. _E.g._, "CHECK_FALSE( !true )", should be "CHECK_FALSE( true )". Fixed.